requirements.txt
openai
ragie
python-dotenv